% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rda.R
\name{load_rda}
\alias{load_rda}
\title{Load objects from inside a .rda file}
\usage{
load_rda(file, envir = parent.frame(), verbose = FALSE, objects = NULL)
}
\arguments{
\item{file}{a (readable binary-mode) connection or a character string giving
the name of the file to load (when tilde expansion is done).}

\item{envir}{the environment where the data should be loaded, by default the
calling environment of the function}

\item{verbose}{should item names be printed during loading? This argument is
used only in the call make to \link{load} function, and so all
object names will be printed even if \code{objects} is specified}

\item{objects}{A vector of character strings naming the objects to load. If
the vector is un-named the objects will be loaded with the names they are
stored under in the .rda file. If the vector is named, the objects will be
renamed as they are loaded.}
}
\description{
Load a subset of objects from a .rda \code{file} into an environment.
If no objects are specified, this function is simply a transparent call
to \link{load}. If objects are specified, only those objects will be loaded
from the .rda file.
}
\section{Side Effects}{

The loaded objects are present in the frame specified in \code{envir}, by
default the calling environment.
}

